Parse Bundle URL too, fixes
